pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openquant"
version = "0.2.0"
description = "AI 半自动量化交易系统"
requires-python = ">=3.9"
dependencies = [
    "akshare>=1.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "jupyter>=1.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
